requirements.txt
defusedxml==0.5.0
diff-match-patch==20181111
Django==1.11.16
django-filter==2.0.0
django-import-export==1.1.0
djangorestframework==3.9.0
et-xmlfile==1.0.1
jdcal==1.4
Markdown==3.0.1
odfpy==1.4.0
openpyxl==2.5.12
Pillow==5.3.0
pkg-resources==0.0.0
psycopg2==2.7.6.1
psycopg2-binary==2.7.6.1
pytz==2018.7
PyYAML==3.13
tablib==0.12.1
unicodecsv==0.14.1
xlrd==1.1.0
xlwt==1.3.0
